add more checks when deleting

+++ codalab/migration.py
@@ -67,15 +67,15 @@ class Migration:
 
         # This file is used to log those bundles's loation that has been changed in database.
         self.change_db_records_file = "change_db_records.txt"
-        self.logger = self.get_logger() 
+        # self.logger = self.get_logger() 
 
-    def get_logger():
+    def get_logger(self):
         """
         Create a logger to log the migration process.
         """
         logger = logging.getLogger(__name__)
         logger.setLevel(logging.INFO)
-        handler = logging.FileHandler(logger.log_file)
+        handler = logging.FileHandler("migration.log")
         handler.setLevel(logging.INFO)
         formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
         handler.setFormatter(formatter)
@@ -110,18 +110,19 @@ class Migration:
 
     def get_bundle_info(self, bundle_uuid, bundle_location):
         target = BundleTarget(bundle_uuid, subpath='')
-        self.logger.info(f"[migration] {target}")
+        logging.info(f"[migration] {target}")
         try:
             info = download_util.get_target_info(bundle_location, target, depth=0)
-            self.logger.info(f"[migration] {info}")
+            logging.info(f"[migration] {info}")
         except Exception as e:
-            self.logger.info(f"[migration] Error: {str(e)}")
+            logging.info(f"[migration] Error: {str(e)}")
             raise e
 
         return info
 
     def upload_to_azure_blob(self, bundle_uuid, bundle_location, is_dir=False):
         # generate target bundle path
+        logging.info(f"[migration] Uploading bundle {bundle_uuid} to Azure storage")
         file_name = "contents.tar.gz" if is_dir else "contents.gz"
         target_location = f"{self.target_store_url}/{bundle_uuid}/{file_name}"
 
@@ -146,7 +147,7 @@ class Migration:
             source_ext = ''
             unpack = False
 
-        self.logger.info("[migration] Uploading from %s to Azure Blob Storage %s", bundle_location, target_location)
+        logging.info("[migration] Uploading from %s to Azure Blob Storage %s", bundle_location, target_location)
         # Upload file content and generate index file
         uploader.write_fileobj(source_ext, source_fileobj, target_location, unpack_archive=unpack)
 
@@ -157,8 +158,8 @@ class Migration:
         Change the bundle location in the database
         ATTENTION: this function will modify codalab
         """
-        self.logger.info(f"[migration] Modifying bundle info {bundle_uuid} in database")
-        
+        logging.info(f"[migration] Modifying bundle info {bundle_uuid} in database")
+
         original_location = self.get_bundle_location(bundle_uuid)
 
         # add bundle location: Add bundle location to database
@@ -184,7 +185,8 @@ class Migration:
         )
         
         # write change database record to a file
-        with open(self.change_db_records_file, 'a') as f:
+        path = os.path.join(self.codalab_manager.codalab_home, self.change_db_records_file)
+        with open(path, 'a') as f:
             f.write(f"{bundle_uuid},{original_location},{new_location}\n")
 
     def sanity_check(self, bundle_uuid, bundle_location, bundle_info, is_dir):
@@ -212,6 +214,7 @@ class Migration:
         Delete original bundle from local disk
         """
         if os.path.exists(bundle_location):
+            logging.info(f"[migration] Deleting original bundle {bundle_uuid} from local disk path {bundle_location}")
             deleted_size = path_util.get_path_size(bundle_location)
             bundle_user_id = self.bundle_manager._model.get_bundle_owner_ids([bundle_uuid])[
                 bundle_uuid
@@ -229,16 +232,21 @@ class Migration:
         """
         Delete all the original bundle that has been uploaded to Azure Blob Storage and changed location in database
         """
-        with open(self.change_db_records_file, "r") as f:
+        path = os.path.join(self.codalab_manager.codalab_home, self.change_db_records_file)
+        if not os.path.exists(path):
+            return
+        with open(path, "r+") as f:
             lines = f.readlines()
-            f.seek(0)
             for line in lines:
-                bundle_uuid, origin_bundle_location, _ = line.split()
+                bundle_uuid, origin_bundle_location, _ = line.split(",")
+                if not self.get_bundle_location(bundle_uuid).startswith(StorageURLScheme.AZURE_BLOB_STORAGE.value):
+                    logging.info(f"Bundle {bundle_uuid} info in database is not properly updated")
+                    raise Exception(f"Bundle {bundle_uuid} info in database is not properly updated")
                 try:
                     self.delete_original_bundle_by_uuid(bundle_uuid, origin_bundle_location)
                 except Exception as e:
                     # If the bundle is not deleted, save the information in the file
-                    self.logger.error(f"[migration] Delete Original Bundle Error: {str(e)}")
+                    logging.error(f"[migration] Delete Original Bundle Error: {str(e)}")
                     f.write(line)
             f.truncate()
                 
@@ -284,10 +292,10 @@ if __name__ == '__main__':
         bundle = migration.get_bundle(bundle_uuid)
 
         # TODO: change this to allow migration of run bundles
-        if bundle.bundle_type != 'dataset' or bundle.state != 'ready':
+        if bundle.state != 'ready':
             # only migrate uploaded bundle, and the bundle state needs to be ready
             continue
-
+            
         # Uploaded bundles does not need has dependencies
         assert len(bundle.dependencies) == 0
